Add files via upload
fixed the save source from using a cached copy. it now gets the script file name a loads it with file open().
added protection for saving nan and inf values to the results data table
turned off protection in the json exporter to dissallow nan and inf values - this is more standards complaint.